Fix 525 findings from full-codebase review across all subsystems
Resolves every confirmed finding from the from-scratch review
(currentfailures.txt): 23 critical, 158 high, 158 medium, 186 low,
spanning autograd, ops/core, nn, models, distributed, jit/lazy,
sparse/quant/nested, io/onnx/serving/export/data, python bindings,
and all backends (cpu/cuda/rocm/oneapi/vulkan).

Highlights:
- Security: bounds/shape/size validation on all deserialization paths
  (dist_checkpoint, lite TLV, nn serialize/checkpoint) closing heap
  overflows from untrusted input.
- Autograd: stop grad_fn severing in vmap/func, detection losses
  (mask_rcnn, mask_head, roi_head), models (albert/vit/t5); per-output
  grad_fn slots for qr/eigh/eig; correct JVP tangents.
- Concurrency: fix ZeRO recursive-lock self-deadlock; oneAPI/CUDA/ROCm
  sparse_trsv SIMT spin-wait deadlock -> sequential solve.
- Backends: reduction-partials dropped beyond one workgroup (cuda dot,
  oneapi prod/norm); ROCm norm mean/var broadcast; rocm argsort dim;
  vulkan adaptive pool batch dim (+ new f16/bf16/f64 shaders).
- Quantization: static quant now bakes calibrated activation qparams
  and strips FakeQuantize modules.
- JIT: interpreter honors fused_relu/fused_bias markers; extended
  codegen compiles the generated kernel.

Verification: full default (CPU) build passes (all 1815 targets);
CUDA/ROCm/Vulkan backends build clean; oneAPI changed files syntax-clean.
Affected unit tests pass; added regression tests (data augmentation,
ZeRO deadlock, JIT fusion, Mask R-CNN gradient flow, static-quant
qparams, lite scalar numel).

+++ include/tenzor/io/stb/stb_image_write.h
@@ -467,7 +467,9 @@ static void stbiw__write_pixels(stbi__write_context *s, int rgb_dir, int vdir, i
 
    for (; j != j_end; j += vdir) {
       for (i=0; i < x; ++i) {
-         unsigned char *d = (unsigned char *) data + (j*x+i)*comp;
+         // Promote to size_t before multiplying so the source-pixel offset does
+         // not overflow 32-bit int for large images (heap OOB read otherwise).
+         unsigned char *d = (unsigned char *) data + ((size_t)j*(size_t)x+(size_t)i)*(size_t)comp;
          stbiw__write_pixel(s, rgb_dir, comp, write_alpha, expand_mono, d);
       }
       stbiw__write_flush(s);
@@ -557,21 +559,22 @@ static int stbi_write_tga_core(stbi__write_context *s, int x, int y, int comp, v
          jdir = -1;
       }
       for (; j != jend; j += jdir) {
-         unsigned char *row = (unsigned char *) data + j * x * comp;
+         // size_t offsets to avoid 32-bit int overflow on large TGA images.
+         unsigned char *row = (unsigned char *) data + (size_t)j * (size_t)x * (size_t)comp;
          int len;
 
          for (i = 0; i < x; i += len) {
-            unsigned char *begin = row + i * comp;
+            unsigned char *begin = row + (size_t)i * (size_t)comp;
             int diff = 1;
             len = 1;
 
             if (i < x - 1) {
                ++len;
-               diff = memcmp(begin, row + (i + 1) * comp, comp);
+               diff = memcmp(begin, row + (size_t)(i + 1) * (size_t)comp, comp);
                if (diff) {
                   const unsigned char *prev = begin;
                   for (k = i + 2; k < x && len < 128; ++k) {
-                     if (memcmp(prev, row + k * comp, comp)) {
+                     if (memcmp(prev, row + (size_t)k * (size_t)comp, comp)) {
                         prev += comp;
                         ++len;
                      } else {
